Analyzer/requirements.txt
streamlit
pdfplumber
python-docx
fpdf2
Pillow
numpy
